+++ Moya/RxSwift/Moya+RxSwift.swift
@@ -4,9 +4,6 @@ import Alamofire
 
 /// Subclass of MoyaProvider that returns Observable instances when requests are made. Much better than using completion closures.
 public class RxMoyaProvider<T where T: MoyaTarget>: MoyaProvider<T> {
-    /// Current requests that have not completed or errored yet.
-    /// Note: Do not access this directly. It is public only for unit-testing purposes (sigh).
-    public var inflightRequests = Dictionary<Endpoint<T>, Observable<MoyaResponse>>()
 
     /// Initializes a reactive provider.
     override public init(endpointClosure: MoyaEndpointsClosure = MoyaProvider.DefaultEndpointMapping, endpointResolver: MoyaEndpointResolution = MoyaProvider.DefaultEndpointResolution, stubBehavior: MoyaStubbedBehavior = MoyaProvider.NoStubbingBehavior, networkActivityClosure: Moya.NetworkActivityClosure? = nil, manager: Manager = Alamofire.Manager.sharedInstance) {
@@ -15,12 +12,8 @@ public class RxMoyaProvider<T where T: MoyaTarget>: MoyaProvider<T> {
 
     /// Designated request-making method.
     public func request(token: T) -> Observable<MoyaResponse> {
-        let endpoint = self.endpoint(token)
 
         return deferred { [weak self] () -> Observable<MoyaResponse> in
-            if let existingObservable = self!.inflightRequests[endpoint] {
-                return existingObservable
-            }
 
             let observable: Observable<MoyaResponse> =  AnonymousObservable { observer in
                 let cancellableToken = self?.request(token) { (data, statusCode, response, error) -> () in
@@ -35,19 +28,8 @@ public class RxMoyaProvider<T where T: MoyaTarget>: MoyaProvider<T> {
                 }
 
                 return AnonymousDisposable {
-                    if let weakSelf = self {
-                        objc_sync_enter(weakSelf)
-                        weakSelf.inflightRequests[endpoint] = nil
-                        cancellableToken?.cancel()
-                        objc_sync_exit(weakSelf)
-                    }
+                    cancellableToken?.cancel()
                 }
-            }
-            
-            if let weakSelf = self {
-                objc_sync_enter(weakSelf)
-                weakSelf.inflightRequests[endpoint] = observable
-                objc_sync_exit(weakSelf)
             }
 
             return observable